ai-service/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
reportlab==4.0.7